[GREEN] Implemented the docx extraction function and make the factory pattern for the design pattern
